requirements.txt
Flask~=3.1.0
fsspec~=2025.3.2
gcsfs~=2025.3.2
google-cloud-storage~=3.1.0
matplotlib~=3.10.1
msgpack~=1.1.0
moviepy~=1.0.3
numpy~=2.2.3
openai~=1.75.0
opencv-python~=4.11.0.86
pandas~=2.2.3
pillow~=10.4.0
PyYAML~=6.0.2
psycopg2-binary~=2.9.10
requests~=2.32.3
simplejpeg~=1.8.2
SQLAlchemy~=2.0.38
sqlitedict~=2.1.0
tqdm~=4.67.1
websockets~=15.0.1

# For development
black~=25.1.0
mypy~=1.15.0
sqlalchemy2-stubs~=0.0.2a38
